Complete UI Service Overhaul
changed how the ui service manages interface plugins. It now longer handles application service events and forwards them to the indivual UIs. Now it is the responsibility of the plugin to subscribe to the application service events that it needs for its particular implementation. The UI service does still manage UI connections and will report any non-xpanel interface errors to all plugins that implement the IErrorInterface interface.

This change is necessary as the requirements of the framework become more complex and unwieldy for build-in behavior management.